requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
beautifulsoup4==4.12.2
requests==2.31.0
mysql-connector-python==8.2.0
google-generativeai==0.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
